refactor(worker): apply review feedback
* Switch worker startup to worker.Run(interruptCh): a small goroutine forwards
  ctx.Done() onto the SDK's interrupt channel, replacing the implicit
  Start + defer Stop + <-ctx.Done() chain with a single explicit blocking
  call.

* Split media.go into config.go (constants, MediaWorkflowConfig, FinalizeMode/
  FinalizeInput), workflow.go (MediaWorkflow + stepError), and activities.go
  (Activities struct, Probe/DetectCrop/Transcode/Finalize and their helpers).
  The workflow / activity boundary now matches the determinism boundary of
  the Temporal SDK.

* Replace the shared mutable failedStep variable with a stepError wrapper:
  each activity-call site returns &stepError{step, err}, and the failure-
  webhook defer extracts the pair via errors.As. The wrapped form
  ("step: message") also surfaces directly in the Temporal Web UI.

* Drop per-call err names (probeErr, cropErr, tErr, finErr) in favour of a
  single scoped err per if-statement.
